Re-organize the structure of procfs
